Loosening packaging trimming tree expansion
Consider two solution constraints added in the following order:
  b1 -> (c<3)
  a1 -> (c<4)

And the following versions of c:
  c2 -> e
  c3 -> f

If we expand only the dependencies of b1 that match its constraint, only e gets
added to the set of reachable packages. This causes the termination condtion of
the tree expansion to trigger when adding a1's dependency on c, believing that
it has already see package c and thus returning.