fixes berkshelf trivial solve failure bug
the probelm here is that calling Array#to_s winds up calling
the #inspect method on all the contents, not the #to_s method.

turns out that with berkshelf calling #inspect winds up walking
some instance variables that have the whole 'universe' of cookbooks
and the first line this patches takes a very long time to even
set the first key value.

by rolling out our own iterator over the array and calling to_s
we avoid descending into instance variables that include the
universe and this completes.

should also improve the cpu processing of exceptions on orgs with
large cookbook collections on EC as well.